chore: prepare Release 41 for release testing
Squashed commit of the following:

commit 1a50478c04228ade7e15341e6e88fdb37af91b81
Author: Jan Starzak <jan.starzak@gmail.com>
Date:   Mon Mar 28 13:07:02 2022 +0200

    chore: dep resolutions

commit 93c4aa77aeade1031638d49bba99e8b322f622dd
Author: Jan Starzak <jan.starzak@gmail.com>
Date:   Tue Mar 22 13:32:38 2022 +0100

    chore: prepare Release 41 for release testing
