+++ meteor/server/migration/migrations.ts
@@ -117,6 +117,9 @@ addSteps1_39_0()
 import { addSteps as addSteps1_40_0 } from './1_40_0'
 addSteps1_40_0()
 
+import { addSteps as addSteps1_41_0 } from './1_41_0'
+addSteps1_41_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
